+++ lib/aws/rails/railtie.rb
@@ -80,11 +80,9 @@ module Aws
     # For example, S3's put_object has an event name of: put_object.S3.aws
     def self.instrument_sdk_operations
       Aws.constants.each do |c|
-        next if Aws.autoload?(c)
-
         m = Aws.const_get(c)
         if m.is_a?(Module) && m.const_defined?(:Client) &&
-           m.const_get(:Client).superclass == Seahorse::Client::Base
+           (client = m.const_get(:Client)) && client.superclass == Seahorse::Client::Base
           m.const_get(:Client).add_plugin(Aws::Rails::Notifications)
         end
       end